requirements-server.txt
redis==7.2.0
fastapi==0.129.0
pydantic==2.12.5
uvicorn==0.41.0
python-dotenv>=1.0
colorlog==6.10.1